+++ lib/pack_core/global_states/app/app.dart
@@ -1,0 +1,1 @@
+export './app_settings_notifier.dart';