+++ packages/wbw_core/lib/src/localization/localization.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/widgets.dart';
+import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:intl/intl.dart';
+import 'package:intl/locale.dart' as locale;
+
+part './language.dart';
+part 'localization.freezed.dart';
+part 'localization.g.dart';
+part 'named_locale.dart';