+++ packages/wbw_core/lib/wbw_core.dart
@@ -2,6 +2,7 @@ library wbw_core;
 
 export 'src/extensions/extensions.dart';
 export 'src/hooks/hooks.dart';
+export 'src/localization/localization.dart';
 export 'src/models/models.dart';
 export 'src/router/router.dart';
 export 'src/services/services.dart';