+++ CHANGELOG.md
@@ -14,6 +14,16 @@
 
 ## Next
 
+## October 12th 2022
+
+- Added Credits dialog
+- Added Icons for Black & White landscape and character
+- Completed translation for previous game languages (English, Russian, Italian (very approximate))
+
+## October 11th 2022
+
+- New: restored an option to switch between languages.
+
 ## October 10th 2022
 
 - Restored an option to add new word to local dictionary.